pyproject.toml
[project]
name = "DDS"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django~=5.2.4",
    "django-admin-rangefilter>=0.13.3",
    "django-autocomplete-light>=3.12.1",
    "django-environ~=0.12.0",
    "django-filter~=25.1",
    "django-silk~=5.4.0",
    "djangorestframework~=3.16.0",
    "djangorestframework-simplejwt~=5.5.1",
    "drf-spectacular~=0.28.0",
    "drf-standardized-errors~=0.15.0",
    "drf-standardized-response~=1.1.3",
    "ruff~=0.12.5",
]

[tool.ruff]
exclude = [
    ".cache/*",
    ".venv",
    "venv",
    "migrations",
    "manage.py",
    "__init__.py",
]

line-length   = 120
indent-width  = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "I",
    "Q",
    "TID",
    "ANN",
    "DJ",
    "T20",
    "B",
    "SIM",
    "ISC002",
    "ISC003",
]
ignore = [
    "ANN401",
    "B904",
    "DJ008",
    "DJ001",
]
preview = true

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint.isort]
split-on-trailing-comma = true
